requirements.txt
beautifulsoup4
discord.py
fastapi[standard]
psycopg
pydantic
python-dotenv
requests